Move postfix builder to a separate file
