+++ src/orange/orangeinp/CsgTreeUtils.cc
@@ -16,7 +16,6 @@
 
 #include "detail/InfixStringBuilder.hh"
 #include "detail/NodeReplacementInserter.hh"
-#include "detail/PostfixLogicBuilderImpl.hh"
 
 namespace celeritas
 {
@@ -138,53 +137,6 @@ void simplify(CsgTree* tree, NodeId start)
     std::sort(result.begin(), result.end());
     CELER_ENSURE(std::unique(result.begin(), result.end()) == result.end());
     return result;
-}
-
-//---------------------------------------------------------------------------//
-/*!
- * Convert a single node to postfix notation.
- *
- * The per-node local surfaces (faces) are sorted in ascending order of ID, not
- * of access, since they're always evaluated sequentially rather than as part
- * of the logic evaluation itself.
- */
-[[nodiscard]] auto PostfixLogicBuilder::operator()(NodeId n) const
-    -> result_type
-{
-    CELER_EXPECT(n < tree_.size());
-
-    // Construct logic vector as local surface IDs
-    VecLogic lgc;
-    detail::PostfixLogicBuilderImpl build_impl{tree_, mapping_, &lgc};
-    build_impl(n);
-
-    // Construct sorted vector of faces
-    std::vector<LocalSurfaceId> faces;
-    for (auto const& v : lgc)
-    {
-        if (!logic::is_operator_token(v))
-        {
-            faces.push_back(LocalSurfaceId{v});
-        }
-    }
-
-    // Sort and uniquify the vector
-    std::sort(faces.begin(), faces.end());
-    faces.erase(std::unique(faces.begin(), faces.end()), faces.end());
-
-    // Remap logic
-    for (auto& v : lgc)
-    {
-        if (!logic::is_operator_token(v))
-        {
-            auto iter
-                = find_sorted(faces.begin(), faces.end(), LocalSurfaceId{v});
-            CELER_ASSUME(iter != faces.end());
-            v = iter - faces.begin();
-        }
-    }
-
-    return {std::move(lgc), std::move(faces)};
 }
 
 //---------------------------------------------------------------------------//
