+++ src/orange/orangeinp/detail/PostfixLogicBuilder.cc
@@ -1,23 +1,24 @@
 //----------------------------------*-C++-*----------------------------------//
-// Copyright 2023-2024 UT-Battelle, LLC, and other Celeritas developers.
+// Copyright 2024 UT-Battelle, LLC, and other Celeritas developers.
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 //---------------------------------------------------------------------------//
-//! \file orange/orangeinp/detail/PostfixLogicBuilderImpl.hh
+//! \file orange/orangeinp/detail/PostfixLogicBuilder.cc
 //---------------------------------------------------------------------------//
-#pragma once
+#include "PostfixLogicBuilder.hh"
 
 #include "corecel/cont/VariantUtils.hh"
 #include "corecel/math/Algorithms.hh"
 #include "orange/OrangeTypes.hh"
-
-#include "../CsgTree.hh"
+#include "orange/orangeinp/CsgTree.hh"
 
 namespace celeritas
 {
 namespace orangeinp
 {
 namespace detail
+{
+namespace
 {
 //---------------------------------------------------------------------------//
 /*!
@@ -71,8 +72,6 @@ class PostfixLogicBuilderImpl
     VecLogic* logic_;
 };
 
-//---------------------------------------------------------------------------//
-// INLINE DEFINITIONS
 //---------------------------------------------------------------------------//
 /*!
  * Construct with pointer to the logic expression.
@@ -179,6 +178,55 @@ void PostfixLogicBuilderImpl::operator()(Joined const& n)
         (*this)(*iter++);
         logic_->push_back(n.op);
     }
+}
+
+//---------------------------------------------------------------------------//
+}  // namespace
+
+//---------------------------------------------------------------------------//
+/*!
+ * Convert a single node to postfix notation.
+ *
+ * The per-node local surfaces (faces) are sorted in ascending order of ID, not
+ * of access, since they're always evaluated sequentially rather than as part
+ * of the logic evaluation itself.
+ */
+auto PostfixLogicBuilder::operator()(NodeId n) const -> result_type
+{
+    CELER_EXPECT(n < tree_.size());
+
+    // Construct logic vector as local surface IDs
+    VecLogic lgc;
+    PostfixLogicBuilderImpl build_impl{tree_, mapping_, &lgc};
+    build_impl(n);
+
+    // Construct sorted vector of faces
+    std::vector<LocalSurfaceId> faces;
+    for (auto const& v : lgc)
+    {
+        if (!logic::is_operator_token(v))
+        {
+            faces.push_back(LocalSurfaceId{v});
+        }
+    }
+
+    // Sort and uniquify the vector
+    std::sort(faces.begin(), faces.end());
+    faces.erase(std::unique(faces.begin(), faces.end()), faces.end());
+
+    // Remap logic
+    for (auto& v : lgc)
+    {
+        if (!logic::is_operator_token(v))
+        {
+            auto iter
+                = find_sorted(faces.begin(), faces.end(), LocalSurfaceId{v});
+            CELER_ASSUME(iter != faces.end());
+            v = iter - faces.begin();
+        }
+    }
+
+    return {std::move(faces), std::move(lgc)};
 }
 
 //---------------------------------------------------------------------------//
